Make new Models.Accounts namespace and move relevant classes into it
