+++ WeddingWebsite/Components/Admin/AccountList.razor
@@ -1,6 +1,6 @@
 @using WeddingWebsite.Data
 @using WeddingWebsite.Data.Models
-@using WeddingWebsite.Models
+@using WeddingWebsite.Models.Accounts
 @using WeddingWebsite.Services
 
 @inject IAdminService AdminService
