+++ WeddingWebsite/Components/Pages/Admin/ManageGuest.razor
@@ -3,7 +3,7 @@
 @using Microsoft.AspNetCore.Identity
 @using WeddingWebsite.Components.Layouts
 @using WeddingWebsite.Data
-@using WeddingWebsite.Models
+@using WeddingWebsite.Models.Accounts
 @using WeddingWebsite.Services
 
 @layout SimpleLayout
